+++ tests/test_main.py
@@ -1,3 +1,4 @@
+import unittest
 from unittest import mock
 
 import pytest
@@ -33,30 +34,38 @@ class TestNonWorkspace:
             workspace_mock.id = constants.WORKSPACE_ID
             yield
 
-    def test_get_webhook_events(self, requests_mock):
-        url_webhook_events = f"{constants.API_HOST}/webhooks/events"
-        events = ["some-event"]
-        requests_mock.get(
-            url=url_webhook_events,
-            json={
-                "data": events,
-                "error": {},
-            },
-        )
+    @unittest.mock.patch("up42.webhooks.Webhooks")
+    def test_should_get_webhook_events(self, webhooks: mock.MagicMock, auth_mock):
+        events = mock.MagicMock()
+        webhooks.return_value.get_webhook_events.return_value = events
         assert main.get_webhook_events() == events
+        webhooks.assert_called_with(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)
 
+    @unittest.mock.patch("up42.webhooks.Webhooks")
     @pytest.mark.parametrize("return_json", [False, True])
-    def test_get_webhooks(self, webhooks_mock, return_json):
-        webhooks = main.get_webhooks(return_json=return_json)
-        expected = webhooks_mock.get_webhooks(return_json=return_json)
-        if return_json:
-            assert webhooks == expected
-        else:
-            for hook, expected_hook in zip(webhooks, expected):
-                assert hook.webhook_id == expected_hook.webhook_id
-                assert hook._info == expected_hook._info  # pylint: disable=protected-access
+    def test_should_get_webhooks(self, webhooks: mock.MagicMock, auth_mock, return_json):
+        hooks = mock.MagicMock()
+        webhooks.return_value.get_webhooks.return_value = hooks
+        assert main.get_webhooks(return_json=return_json) == hooks
+        webhooks.assert_called_with(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)
+        webhooks.return_value.get_webhooks.assert_called_with(return_json=return_json)
 
-    def test_get_credits_balance(self, requests_mock):
+    @unittest.mock.patch("up42.webhooks.Webhooks")
+    def test_should_create_webhook(self, webhooks: mock.MagicMock, auth_mock):
+        name = "name"
+        url = "url"
+        events = ["event"]
+        active = True
+        secret = "secret"
+        webhook = mock.MagicMock()
+        webhooks.return_value.create_webhook.return_value = webhook
+        assert webhook == main.create_webhook(name, url, events, active, secret)
+        webhooks.assert_called_with(auth=auth_mock, workspace_id=constants.WORKSPACE_ID)
+        webhooks.return_value.create_webhook.assert_called_with(
+            name=name, url=url, events=events, active=active, secret=secret
+        )
+
+    def test_should_get_credits_balance(self, requests_mock):
         balance_url = f"{constants.API_HOST}/accounts/me/credits/balance"
         balance = {"balance": 10693}
         requests_mock.get(